requirements.txt
aiogram>=3.4.0
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
werkzeug>=3.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
requests>=2.31.0